6 - extrating a trait
