+++ src/Lookitsatravis/Listify/Listify.php
@@ -1,10 +1,6 @@
 <?php namespace Lookitsatravis\Listify;
 
 use DB, Event, Config, App;
-use Lookitsatravis\Listify\Exceptions\ListifyException;
-use Lookitsatravis\Listify\Exceptions\NullForeignKeyException;
-use Lookitsatravis\Listify\Exceptions\NullScopeException;
-use Lookitsatravis\Listify\Exceptions\InvalidScopeException;
 use Lookitsatravis\Listify\Exceptions\InvalidQueryBuilderException;
 
 /**
@@ -21,6 +17,7 @@ use Lookitsatravis\Listify\Exceptions\InvalidQueryBuilderException;
 
 trait Listify
 {
+    use ScopeCondition;
     /**
      * Array of current config values
      * @var array
@@ -157,49 +154,6 @@ trait Listify
         }
 
         return FALSE;
-    }
-
-    /**
-     * Returns the raw WHERE clause to be used as the Listify scope
-     * @return string
-     */
-    private function scopeCondition()
-    {
-        $theScope = $this->scopeName();
-
-        if($theScope === NULL)
-        {
-            throw new NullScopeException('You cannot pass in a null scope into Listify. It breaks stuff.');
-        }
-
-        if($theScope === $this->defaultScope) {
-            return $theScope;
-        }
-        if (is_string($theScope)) {
-            //Good for you for being brave. Let's hope it'll run in your DB! You sanitized it, right?
-            $this->stringScopeValue = $theScope;
-            return $theScope;
-        }
-        if (! is_object($theScope)) {
-            throw new InvalidScopeException('Listify scope parameter must be a String, an Eloquent BelongsTo object, or a Query Builder object.');
-        }
-        $reflector = new \ReflectionClass($theScope);
-        if ($reflector->getName() == 'Illuminate\Database\Eloquent\Relations\BelongsTo')
-        {
-            $relationshipId = $this->getAttribute($theScope->getForeignKey());
-
-            if ($relationshipId === null) {
-                throw new NullForeignKeyException('The Listify scope is a "belongsTo" relationship, but the foreign key is null.');
-            }
-
-            return $theScope->getForeignKey().' = '.$this->getAttribute($theScope->getForeignKey());
-        }
-
-        if ($reflector->getName() != 'Illuminate\Database\Query\Builder') {
-            throw new InvalidScopeException('Listify scope parameter must be a String, an Eloquent BelongsTo object, or a Query Builder object.');
-        }
-
-        return $this->stringScopeValue = $this->getConditionStringFromQueryBuilder($theScope);
     }
 
     /**
